requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bpemb=0.3.3=pypi_0
ca-certificates=2021.10.8=h5b45459_0
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.11=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pypi_0
conllu=4.4.1=pypi_0
cryptography=36.0.1=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.23=pypi_0
deprecated=1.2.13=pypi_0
doc2text=0.2.4=pypi_0
filelock=3.4.2=pypi_0
flair=0.10=pypi_0
fonttools=4.29.1=pypi_0
ftfy=6.0.3=pypi_0
future=0.18.2=pypi_0
gdown=3.12.2=pypi_0
gensim=4.1.2=pypi_0
huggingface-hub=0.4.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=3.10.1=pypi_0
janome=0.4.1=pypi_0
joblib=1.1.0=pypi_0
kiwisolver=1.3.2=pypi_0
konoha=4.6.5=pypi_0
langdetect=1.0.9=pypi_0
libiconv=1.16=he774522_0
libxml2=2.9.12=h0ad7f3c_0
libxslt=1.1.33=h579f668_0
libzlib=1.2.11=h8ffe710_1013
lxml=4.7.1=py39h4fd7cdf_0
matplotlib=3.5.1=pypi_0
mime=0.1.0=pypi_0
more-itertools=8.8.0=pypi_0
mpld3=0.3=pypi_0
numpy=1.22.1=pypi_0
openssl=1.1.1l=h2bbff1b_0
overrides=3.1.0=pypi_0
packaging=21.3=pypi_0
pdfminer-six=20211012=pypi_0
pdfplumber=0.6.0=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=py39haa95532_0
protobuf=3.19.4=pypi_0
pycparser=2.21=pypi_0
pyparsing=3.0.7=pypi_0
pypdf2=1.26.0=pypi_0
pysocks=1.7.1=pypi_0
pytesseract=0.3.8=pypi_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pypi_0
python-docx=0.8.11=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pyyaml=6.0=pypi_0
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
sacremoses=0.0.47=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
segtok=1.5.11=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=58.0.4=py39haa95532_0
six=1.16.0=pypi_0
smart-open=5.2.1=pypi_0
sqlite=3.37.0=h2bbff1b_0
sqlitedict=1.7.0=pypi_0
tabulate=0.8.9=pypi_0
threadpoolctl=3.1.0=pypi_0
tokenizers=0.10.3=pypi_0
torch=1.10.2=pypi_0
tqdm=4.62.3=pypi_0
transformers=4.15.0=pypi_0
typing-extensions=4.0.1=pypi_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wand=0.6.7=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wikipedia-api=0.5.4=pypi_0
wincertstore=0.2=py39haa95532_2
wrapt=1.13.3=pypi_0
xz=5.2.5=h62dcd97_0
zipp=3.7.0=pypi_0
zlib=1.2.11=h8ffe710_1013
